update : divide all filters in Components.
