+++ src/Pages/ProductPage/Components/FilteredData/SearchProduct.jsx
@@ -1,0 +1,9 @@
+export const SearchProduct = (data, sortBy) => {
+  let filteredList = data;
+  if (sortBy) {
+    return filteredList.filter((item) =>
+      item.title.toLowerCase().match(sortBy.toLowerCase())
+    );
+  }
+  return filteredList;
+};
