+++ src/Pages/ProductPage/Components/FilteredData/index.js
@@ -1,0 +1,5 @@
+export { SearchProduct } from "./SearchProduct.jsx";
+export { SortByPrice } from "./SortByPrice.jsx";
+export { SortByCategory } from "./SortByCategory.jsx";
+export { SortByRating } from "./SortByRating.jsx";
+export { SortByRange } from "./SortByRange.jsx";
